requirements.txt
streamlit
pillow
pytesseract
easyocr
transformers
torch
sentence-transformers
spacy
pdf2image
pdfplumber
scikit-learn
numpy
pandas
openpyxl
